Move calculateValidTo() to time utils
