--- src/cow-react/common/utils/calculateValidTo.ts
@@ -1,12 +1,0 @@
-import { BigNumber } from '@ethersproject/bignumber'
-
-export const MAX_VALID_TO_EPOCH = BigNumber.from('0xFFFFFFFF').toNumber() // Max uint32 (Feb 07 2106 07:28:15 GMT+0100)
-
-export function calculateValidTo(deadline: number): number {
-  // Need the timestamp in seconds
-  const now = Date.now() / 1000
-  // Must be an integer
-  const validTo = Math.floor(deadline + now)
-  // Should not be greater than what the contract supports
-  return Math.min(validTo, MAX_VALID_TO_EPOCH)
-}
