+++ src/cow-react/utils/time.ts
@@ -2,6 +2,8 @@
  * Given a Date obj, remove hours, minutes and seconds, and return the timestamp of it
  * @param date
  */
+import { BigNumber } from '@ethersproject/bignumber'
+
 export function getDateTimestamp(date: Date): number {
   return new Date(date.getFullYear(), date.getMonth(), date.getDate()).getTime()
 }
@@ -26,4 +28,15 @@ export function timeSinceInSeconds(timestamp?: number): number | undefined {
   const timeDelta = Date.now() - timestamp
 
   return Math.floor(timeDelta / 1000)
+}
+
+export const MAX_VALID_TO_EPOCH = BigNumber.from('0xFFFFFFFF').toNumber() // Max uint32 (Feb 07 2106 07:28:15 GMT+0100)
+
+export function calculateValidTo(deadline: number): number {
+  // Need the timestamp in seconds
+  const now = Date.now() / 1000
+  // Must be an integer
+  const validTo = Math.floor(deadline + now)
+  // Should not be greater than what the contract supports
+  return Math.min(validTo, MAX_VALID_TO_EPOCH)
 }
